separated views to multiple views to make it more readable code
