+++ views/API_views.py
@@ -1,0 +1,69 @@
+from django.http import JsonResponse
+from django.views.decorators.csrf import csrf_exempt
+from rest_framework import generics ,status
+from rest_framework.parsers import JSONParser
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+
+from ..models import Task
+from ..serializers import TaskSerializer
+
+
+
+
+@api_view(['GET','POST'])  #list of methods that are allowed are GET and POST .. if delete is not 
+                           # mentioned above ---> means that once delete is choosen it will show 
+                           # error 
+@csrf_exempt
+def task_list_api(request):
+    if request.method == "POST":
+        data = JSONParser().parse(request)
+        serializer = TaskSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return JsonResponse(serializer.data, status = status.HTTP_201_CREATED )
+        return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST )
+    elif request.method == "GET":
+        tasks = Task.objects.all()
+        serializer = TaskSerializer(tasks, many=True)
+        return JsonResponse(serializer.data, safe=False)
+
+
+@api_view(['GET', 'POST', 'PUT', 'DELETE'])
+@csrf_exempt
+def task_detail_api(request, pk):
+    try:
+        task = Task.objects.get(id=pk)
+    except Task.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == "GET":
+        serializer = TaskSerializer(task)
+        return Response(serializer.data)
+    elif request.method == "POST":
+        data = JSONParser().parse(request)
+        serializer = TaskSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == "PUT":
+        data = JSONParser().parse(request)
+        serializer = TaskSerializer(task, data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == "DELETE":
+        task.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+        
+
+class TaskListCreateView(generics.ListCreateAPIView):
+    queryset = Task.objects.all()
+    serializer_class = TaskSerializer
+
+
+class TaskRetrieveUpdateDeleteView(generics.RetrieveUpdateDestroyAPIView):
+    queryset = Task.objects.all()
+    serializer_class = TaskSerializer
