+++ views/__init_.py
@@ -1,0 +1,3 @@
+from .API_views import *
+from .login_views import *
+from .task_views import *
