+++ views/login_views.py
@@ -1,0 +1,70 @@
+from django import forms
+from django.conf import settings
+from django.contrib.auth import login
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.models import User
+from django.contrib.auth.views import LoginView, LogoutView
+from django.core.mail import send_mail
+from django.shortcuts import redirect
+from django.urls import reverse_lazy
+from django.views.generic.edit import FormView
+                                       
+
+class CustomLoginView(LoginView):
+    template_name = 'base/login.html'
+    fields = '__all__'
+    redirect_authenticated_user = True
+    
+    def get_success_url(self):
+        self.request.session['username'] = self.request.user.username
+        return reverse_lazy('tasks')
+
+
+class CustomUserCreationForm(UserCreationForm):
+    email = forms.EmailField(required=True)
+    
+    class Meta:
+        model = User
+        fields = ['username','first_name','last_name', 'email', 'password1']
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields['username'].help_text = ''
+        self.fields['password1'].help_text = ''
+
+
+class SignUp(FormView):
+    template_name = "base/signup.html"
+    form_class = CustomUserCreationForm # we have just inherited from this builtin Form
+    redirect_authenticated_user = True
+    success_url = reverse_lazy('tasks')
+
+    def form_valid(self, form):
+        user = form.save()
+        if user is not None:
+            login(self.request, user)
+            # send confirmation email
+            subject = 'Welcome to My To Do List!'
+            message = 'Thank you for signing up!'
+            from_email = settings.DEFAULT_FROM_EMAIL
+            recipient_list = [user.email]
+            send_mail(subject, message, from_email, recipient_list)
+        return super().form_valid(form)
+
+    def dispatch(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            return redirect(self.success_url)
+        return super().dispatch(request, *args, **kwargs)
+
+
+    def get_form(self, form_class= None):
+        form = super().get_form(form_class)
+        del form.fields['password2']
+        return form
+
+
+
+
+class CustomLogoutView(LogoutView):
+    def get_default_redirect_url(self):
+        return reverse_lazy ('login')
