+++ views/task_views.py
@@ -1,0 +1,58 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.urls import reverse_lazy
+from django.views.generic.detail import DetailView
+from django.views.generic.edit import (CreateView, DeleteView,
+                                       UpdateView)
+from django.views.generic.list import ListView
+from ..models import Task
+
+
+
+class TaskList(LoginRequiredMixin, ListView):
+    model = Task
+    context_object_name = 'tasks'
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['tasks'] = context['tasks'].filter(user=self.request.user)
+        context['count'] = context['tasks'].filter(complete=False).count()
+        context['completed'] = context['tasks'].filter(complete=True)
+        search_input = self.request.GET.get('search-area') or ''
+        if search_input:
+            context['tasks'] = context['tasks'].filter(title__icontains=search_input)
+        context['search_input'] = search_input
+        context['username'] = self.request.session.get('username', 0)
+
+        view_count = self.request.session.get('view_count',0)+1
+        self.request.session['view_count'] = view_count
+        self.request.session.modified = True
+        context['view_count'] = view_count
+        return context
+
+
+
+class TaskDetail(LoginRequiredMixin, DetailView):
+    model = Task
+    context_object_name = 'task'
+    template_name = 'base/task.html'
+
+class TaskCreate(LoginRequiredMixin, CreateView):
+    model = Task
+    fields = ['title','description','complete']
+    success_url = reverse_lazy('tasks')
+
+    def form_valid(self, form): # instead of choosing the user who created the task
+                                # this auto changes it to the logged user without having to 
+                                # take care of changing it by your self everytime . 
+        form.instance.user = self.request.user
+        return super(TaskCreate,self).form_valid(form)
+
+class TaskUpdate(LoginRequiredMixin, UpdateView):
+    model = Task
+    fields = ['title','description','complete']
+    success_url = reverse_lazy('tasks')
+
+class TaskDelete(LoginRequiredMixin, DeleteView):
+    model = Task 
+    context_object_name = 'task'
+    success_url = reverse_lazy ('tasks')
